Move PlatformConfig classes/functions out of arrow_adapter
