+++ libtiledbsoma/src/soma/soma_collection.h
@@ -16,6 +16,7 @@
 
 #include <tiledb/tiledb>
 
+#include "../tiledb_adapter/platform_config.h"
 #include "enums.h"
 #include "soma_dataframe.h"
 #include "soma_dense_ndarray.h"
