+++ libtiledbsoma/src/soma/soma_dataframe.h
@@ -18,6 +18,7 @@
 #include <optional>
 
 #include "../common/soma_column_selection.h"
+#include "../tiledb_adapter/platform_config.h"
 #include "soma_array.h"
 
 namespace tiledbsoma {
