+++ libtiledbsoma/src/soma/soma_geometry_column.h
@@ -27,6 +27,7 @@
 #include <vector>
 
 #include <tiledb/tiledb>
+#include "../tiledb_adapter/platform_config.h"
 #include "soma_column.h"
 #include "soma_coordinates.h"
