+++ libtiledbsoma/src/soma/soma_geometry_dataframe.h
@@ -17,6 +17,7 @@
 #include <filesystem>
 #include <vector>
 
+#include "../tiledb_adapter/platform_config.h"
 #include "soma_array.h"
 #include "soma_coordinates.h"
