+++ libtiledbsoma/src/soma/soma_point_cloud_dataframe.h
@@ -16,6 +16,7 @@
 
 #include <filesystem>
 
+#include "../tiledb_adapter/platform_config.h"
 #include "soma_array.h"
 #include "soma_coordinates.h"
