+++ libtiledbsoma/src/soma/soma_sparse_ndarray.h
@@ -21,6 +21,7 @@
 #include <vector>
 
 #include "../common/soma_column_selection.h"
+#include "../tiledb_adapter/platform_config.h"
 #include "soma_array.h"
 
 namespace tiledbsoma {
