+++ libtiledbsoma/src/tiledbsoma/tiledbsoma
@@ -19,6 +19,7 @@
 #include "tiledbsoma_export.h"
 
 #include "common/soma_column_selection.h"
+#include "tiledb_adapter/platform_config.h"
 #include "utils/arrow_adapter.h"
 #include "utils/common.h"
 #include "utils/transformer.h"
